bikeshed, rename server -> ranch
